Move all files into their new directories, still need to remove expect/actual.
