--- jvm/selfie-lib/src/jsMain/kotlin/com/diffplug/selfie/SnapshotFile.js.kt
@@ -1,26 +1,0 @@
-/*
- * Copyright (C) 2023-2024 DiffPlug
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.diffplug.selfie
-
-actual class LineReader {
-  actual companion object {
-    actual fun forString(content: String): LineReader = TODO()
-    actual fun forBinary(content: ByteArray): LineReader = TODO()
-  }
-  actual fun getLineNumber(): Int = TODO()
-  actual fun readLine(): String? = TODO()
-  actual fun unixNewlines(): Boolean = TODO()
-}
