--- jvm/selfie-lib/src/jsMain/kotlin/com/diffplug/selfie/guts/Literals.js.kt
@@ -1,22 +1,0 @@
-/*
- * Copyright (C) 2024 DiffPlug
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.diffplug.selfie.guts
-
-/**
- * If the user is hip enough to run javascript, they're probably hip enough for multiline string
- * literals.
- */
-internal actual fun jreVersion(): Int = 15
