+++ jvm/selfie-lib/src/main/kotlin/com/diffplug/selfie/Mode.kt
