+++ jvm/selfie-lib/src/main/kotlin/com/diffplug/selfie/PerCharacterEscaper.kt
