+++ jvm/selfie-lib/src/main/kotlin/com/diffplug/selfie/guts/WriteTracker.kt
