+++ jvm/selfie-lib/src/test/kotlin/com/diffplug/selfie/PerCharacterEscaperTest.kt
