+++ jvm/selfie-lib/src/test/kotlin/com/diffplug/selfie/guts/LiteralLongTest.kt
