+++ jvm/selfie-lib/src/test/resources/com/diffplug/selfie/example.ss
