+++ jvm/selfie-lib/src/test/resources/com/diffplug/selfie/scenarios_and_lenses.ss
